fixed small bug in how middle monomers are assigned
Old implementation assumes that the provided sequence string will be "A" -> "Z" with no gaps. This allows one to index naievely into the list returned by monogrp.iter_rdmols(term_only=False). 

This breaks if a user passes a discontinuous string, such as "ACDE". C will be incorrectly mapped to the second monomer, and the off-by-one error will propagate.